learning mode

+++ function/factorial.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 using namespace std;
-unsigned long long int fact(int num);
+long int fact(int num);
 int main(int argc, char const *argv[])
 {
     int number;
@@ -9,9 +9,9 @@ int main(int argc, char const *argv[])
     cout << "factorial is:" << fact(number);
     return 0;
 }
-unsigned long long  int fact(int num)
+long int fact(int num)
 {
-    unsigned long long  int factorial = 1;
+    long  int factorial = 1;
     for (int i = 1; i <= num; i++)
     {
         factorial *= i;
